requirements.txt
pandas
pandas_ta
python-binance
numpy==1.23.5
apscheduler
requests
matplotlib
ipykernel
gym==0.21.0
pydantic==1.10.13
ray[rllib]==2.2.0
torch
pyyaml
ta
aiohttp
raylib-py
flask
gputil
pytest
pyspark
fastapi
uvicorn
confluent-kafka>=1.8.2
azure-functions
azure-storage-blob>=12.10.0
tqdm
joblib